requirement.txt
numpy
pandas
pytorch
matplotlib
pyyaml
scikit-learn
h5py
openseespy